fix(web): unblock CI on new-sidebar-navigation
- Decompose WorkspaceRail.tsx (1599 LOC → 1048 LOC) by extracting
  RailIconButton, useInboxCount, and WorkspaceRailTools +
  WorkspaceRailFooter into sibling modules. Clears the >=1500 LOC
  file-size budget that the lint job enforces.
- Rewrite WorkspaceRail.test.tsx to match the new switcher-modal
  architecture (only the active workspace renders inline; other
  workspaces + create + kebab live inside the modal).
- Trim AgentList.test.tsx of the peek-chevron expectations (the
  chevron was removed; hover/long-press peek path stays).
- Replace ChannelHeader.test.tsx with a smoke test — theme switcher
  and inline search affordances were both removed from the header.
- Update e2e selectors:
  • nex-connect: open Settings via the WorkspaceRail footer instead
    of the (now removed) sidebar-header gear.
  • sidebar-scroll: drop the .sidebar-apps assertions (Tools moved
    to the rail); confirm Settings is reachable via the rail.
  • app-routes: drive Console ↔ Graph swap through the rail's
    More-tools popover and assert active tools via
    `aria-current="page"`.
